requirements.txt
azure-common==1.1.28
azure-core==1.23.0
azure-data-tables==12.3.0
azure-functions==1.9.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.10.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
idna==3.3
isodate==0.6.1
lxml==4.8.0
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.3
pandas==1.4.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
typing-extensions==4.1.1
urllib3==1.26.9
